Rename receivers and services packages to make them consistent with the rest of the packages
