+++ smssync/src/main/AndroidManifest.xml
@@ -204,6 +204,6 @@
                     android:value=".presentation.view.ui.activity.SettingsActivity"/>
         </activity>
         <!-- Services -->
-        <service android:name="org.addhen.smssync.presentation.services.SmsReceiverService"/>
+        <service android:name="org.addhen.smssync.presentation.service.SmsReceiverService"/>
     </application>
 </manifest>
