+++ smssync/src/main/java/org/addhen/smssync/presentation/service/ServiceConstants.java
@@ -15,7 +15,7 @@
  * Ushahidi developers at team@ushahidi.com.
  */
 
-package org.addhen.smssync.presentation.services;
+package org.addhen.smssync.presentation.service;
 
 /**
  * @author Ushahidi Team <team@ushahidi.com>
